requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
pydantic>=2.5

# Google Sheets
gspread>=6.0
google-auth>=2.28

# Browser automation (Step 3 EMR + Step 5 cathlab)
playwright>=1.42

# LLM SDKs — the user installs whichever they'll use
anthropic>=0.40
openai>=1.40
google-genai>=0.3

# Taiwan holidays (used by scheduling rules)
holidays>=0.45
